test: restructure test suite with PSR-4 compliance and shared HttpTestHelpers trait
