+++ backend/tests/Services/ResponseFormatterTest.php
@@ -1,4 +1,6 @@
 <?php
+namespace Tests\Services;
+
 use PHPUnit\Framework\TestCase;
 use App\Services\ResponseFormatter;
