+++ composer.json
@@ -19,7 +19,12 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\":"backend/src/"
+            "App\\": "backend/src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "backend/tests/"
         }
     },
     "scripts": {
